Moved Address model
